tensor dmr test

+++ src/cuda/gemm_tensorcores/tst/tensor_dmr/simpleTensorCoreGEMM.cu
@@ -158,7 +158,8 @@ __global__ void wmma_example(half *a, half *b, float *c, int M, int N, int K, fl
 //  3) Neither A nor B are transposed.
 // Note: This is NOT a high performance example but is for demonstration purposes only
 //       For a high performance code please use the GEMM provided in cuBLAS.
-__global__ void wmma_example_dmr(half *a, half *b, float *c, half *d_sw, int M, int N, int K, float alpha, float beta) {
+
+__global__ void wmma_example_dmr(half *a, half *b, float *c, float *d_sw, int M, int N, int K, float alpha, float beta) {
    // Leading dimensions. Packed with no transpositions.
    int lda = M;
    int ldb = K;
